requirements.txt
opencv-python==4.4.0.42
pip==20.1
utils==1.0.1
zlib==1.2.11
wheel==0.33.4
google==3.0.0
openssl==1.1.1g
cryptography==2.7
py==1.8.0
lxml==4.3.4
matplotlib==3.3.0
numpy==1.18.5
torch==1.6.0
torchvision==0.7.0
yaml==0.1.7
pyyaml==5.1.1
scipy==1.4.1
tqdm==4.32.1
pillow==7.2.0
future==0.17.1
pygame==1.9.6
pandas==0.24.2
pillow==7.2.0